Update index.html
Commented out stringify, that was for my own readability. Too much information was being presented when calling a summoner's match history so I edited it to only show the boolean value of whether the match was a win or loss. Next steps are to write a loop to cycle through all 10 matches if applicable checking for a win or loss.