Update index.html
Added a for loop to go through each match in a given summoner's history. For some reason I can't get the code to write out the results as a list but it works just fine in alert, so I'm guessing it has more to do with a formatting issue on the html side. Will have to look into that later. Maybe put some pictures with it :D